requirements.txt
ollama>=0.1.14
rich>=13.0.0
python-dotenv>=1.0.0